services/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.32
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.1
psycopg2-binary==2.9.9